requirements.txt
pandas
numpy
scikit-learn
lightgbm
xgboost
scipy
seaborn
cloudpickle
cachetools
imbalanced-learn
optuna
tensorflow
matplotlib
boto3
snowflake-connector-python[pandas]
snowflake-snowpark-python[pandas]
python-dotenv